added ascii banner removed tendermint subcommand
