+++ cmd/rollappd/start.go
@@ -13,7 +13,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/server"
 	tmcmd "github.com/tendermint/tendermint/cmd/tendermint/commands"
-	"github.com/tendermint/tendermint/libs/cli"
 	"github.com/tendermint/tendermint/node"
 	"github.com/tendermint/tendermint/p2p"
 	"github.com/tendermint/tendermint/proxy"
@@ -28,7 +27,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/server/types"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 
-	"github.com/cosmos/cosmos-sdk/version"
 	"github.com/spf13/cobra"
 
 	dymintconf "github.com/dymensionxyz/dymint/config"
@@ -441,43 +439,4 @@ func startInProcess(ctx *server.Context, clientCtx client.Context, appCreator ty
 
 	// wait for signal capture and gracefully return
 	return WaitForQuitSignals()
-}
-
-// add Rollapp commands
-func AddRollappCommands(rootCmd *cobra.Command, defaultNodeHome string, appCreator types.AppCreator, appExport types.AppExporter, addStartFlags types.ModuleInitFlags) {
-	tendermintCmd := &cobra.Command{
-		Use:   "tendermint",
-		Short: "Tendermint subcommands",
-	}
-
-	tendermintCmd.AddCommand(
-		server.VersionCmd(),
-	)
-
-	dymintCmd := &cobra.Command{
-		Use:   "dymint",
-		Short: "Dymint subcommands",
-	}
-
-	dymintCmd.AddCommand(
-		ShowSequencer(),
-		ShowNodeIDCmd(),
-		ResetAll(),
-		InitFiles(),
-		tmcmd.ResetStateCmd,
-	)
-
-	dymintCmd.PersistentFlags().StringP(cli.HomeFlag, "", defaultNodeHome, "directory for config and data")
-
-	startCmd := StartCmd(appCreator, defaultNodeHome)
-	addStartFlags(startCmd)
-
-	rootCmd.AddCommand(
-		startCmd,
-		dymintCmd,
-		tendermintCmd,
-		server.ExportCmd(appExport, defaultNodeHome),
-		version.NewVersionCommand(),
-		server.NewRollbackCmd(appCreator, defaultNodeHome),
-	)
 }
